backend/requirements.txt
django
djangorestframework
django-cors-headers
whitenoise[brotli]
dj-database-url
psycopg2-binary
gunicorn
python-dotenv
asgiref
sqlparse
tzdata
